Move output sorting code to a new file
